the get_account method moved to the chain_plugin #406
