+++ plugins/chain_plugin/include/eosio/chain_plugin/chain_plugin_params.hpp
@@ -3,10 +3,18 @@
 #include <boost/container/flat_set.hpp>
 
 #include <eosio/chain/types.hpp>
+#include <eosio/chain/symbol.hpp>
 
 #include <fc/optional.hpp>
 
 namespace eosio {
+
+    struct get_account_params {
+       chain::name                 account_name;
+       fc::optional<chain::symbol> expected_core_symbol;
+    };
+
+
     struct get_info_params{};
 
     struct get_block_params {
@@ -25,6 +33,7 @@ namespace eosio {
 
 }
 
+FC_REFLECT( eosio::get_account_params, (account_name)(expected_core_symbol) )
 FC_REFLECT_EMPTY(eosio::get_info_params )
 FC_REFLECT(eosio::get_block_params, (block_num_or_id))
 FC_REFLECT(eosio::get_block_header_state_params, (block_num_or_id))
